+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+
+- Moved `SendEmailBatchResponse` from `email` to `batch` module
+- Made `LocatedError`, `DebugResult` and `CLIENT` (all used for testing internals) private (they
+  were accidentally public before, very unlikely they were used by anyone)
 
 ## Added